clean up useless stuff
