--- src/lens/config.js
@@ -1,7 +1,0 @@
-import { compose, lensProp } from 'ramda';
-
-const configLens = lensProp('config');
-const featuresLens = lensProp('features');
-const configFeaturesLens = compose(configLens, featuresLens);
-
-export { configLens, featuresLens, configFeaturesLens };
