--- src/util/benchmark.js
@@ -1,19 +1,0 @@
-import { always, compose } from 'ramda';
-import { debugIt } from './debug';
-
-export const hrTimeToMs = ([seconds, nanoseconds]) =>
-  Number(seconds * 1000 + nanoseconds / 10 ** 6);
-
-export const measureTime = () =>
-  compose(hrTimeToMs, process.hrtime, always(process.hrtime()));
-
-export const debugItTime = () => {
-  const measure = measureTime();
-
-  return () => {
-    const ms = measure();
-    debugIt(`${ms}ms`);
-
-    return ms;
-  };
-};
