--- src/util/reader.js
@@ -1,4 +1,0 @@
-import { apply, compose, identity, unapply } from 'ramda';
-import { Reader } from 'monet';
-
-export const callReader = compose(Reader, unapply, apply, identity);
